Refactor terminal component
