+++ components/terminal/terminal.tsx
@@ -16,101 +16,21 @@ import {
 
 type TermHandle = ElementRef<typeof Term>;
 
+const TerminalHeader = () => (
+  <div className="shrink-0 sticky inset-x-0 top-0 bg-white/50 backdrop-blur-md grid grid-cols-3 px-4 py-3 border-b">
+    <div className="flex items-center gap-2">
+      <div className="rounded-full bg-red-500 w-3.5 h-3.5"></div>
+      <div className="rounded-full bg-yellow-500 w-3.5 h-3.5"></div>
+      <div className="rounded-full bg-green-500 w-3.5 h-3.5"></div>
+    </div>
+    <div className="text-center">bash</div>
+  </div>
+);
+
 export default function Terminal() {
   const containerRef = useRef<HTMLDivElement>(null);
   const term = useRef<TermHandle>(null);
-  const { history, pushToHistory, resetHistory } = useTerminal();
-
-  const commands: TerminalCommands = useMemo(
-    () => ({
-      empty: async () => {
-        await pushToHistory({ command: "" });
-      },
-      help: async () => {
-        const item: TerminalHistoryItem = {
-          command: "help",
-          response: (
-            <>
-              <p>ls [-l]: list directory contents</p>
-              <p>cat: print files</p>
-              <p>clear: clear output</p>
-              <p>sh: run an executable file</p>
-            </>
-          ),
-        };
-        await pushToHistory(item);
-      },
-      "not-found": (command: string) => {
-        pushToHistory({
-          command,
-          response: `Command not found : '${command}'. Type 'help' for available commands.`,
-        });
-      },
-      clear: async () => {
-        await resetHistory();
-      },
-      ls: async () => {
-        const item: TerminalHistoryItem = {
-          command: "ls",
-          response: (
-            <div className="space-x-8">
-              {Array.from(files.keys()).map((fileName) => (
-                <span key={fileName}>{fileName}</span>
-              ))}
-            </div>
-          ),
-        };
-        await pushToHistory(item);
-      },
-      cat: async (fileName: string) => {
-        const file = files.get(fileName);
-        let response;
-        if (file) {
-          if (file.type === "text") {
-            response = file.content;
-          } else {
-            response = "cat: cannot read executable file";
-          }
-        } else {
-          response = `cat: ${fileName}: No such file`;
-        }
-        const item: TerminalHistoryItem = {
-          command: `cat ${fileName}`,
-          response,
-        };
-        await pushToHistory(item);
-      },
-      sh: async (fileName: string) => {
-        const file = files.get(fileName);
-        let response;
-        if (file) {
-          if (file.type === "executable") {
-            response = file.content;
-          } else {
-            response = "sh: cannot execute this file type";
-          }
-        } else {
-          response = `sh: ${fileName}: No such file`;
-        }
-        const item: TerminalHistoryItem = {
-          command: `sh ${fileName}`,
-          response,
-        };
-        await pushToHistory(item);
-      },
-    }),
-    [pushToHistory, resetHistory]
-  );
-
-  // add initial history
-  useEffect(() => {
-    pushToHistory({
-      command: "cat welcome.txt",
-      response:
-        "Hey Folks! Welcome to my personal website where you can discover more about me, my interest for web development and where you can find me when I'm not working.",
-    });
-    commands["help"]();
-  }, [commands, pushToHistory]);
+  const { history, resetHistory, commands } = useTerminal();
 
   const onKeyDown: KeyboardEventHandler = useCallback(
     (event) => {
@@ -122,6 +42,7 @@ export default function Terminal() {
     [resetHistory]
   );
 
+  // focus and scroll to terminal prompt input field
   const focusInput = useCallback(() => {
     // focus only when no text is selected
     if (window.getSelection().isCollapsed) {
@@ -145,17 +66,6 @@ export default function Terminal() {
       focusInput();
     },
     [commands, focusInput]
-  );
-
-  const TerminalHeader = () => (
-    <div className="shrink-0 sticky inset-x-0 top-0 bg-white/50 backdrop-blur-md grid grid-cols-3 px-4 py-3 border-b">
-      <div className="flex items-center gap-2">
-        <div className="rounded-full bg-red-500 w-3.5 h-3.5"></div>
-        <div className="rounded-full bg-yellow-500 w-3.5 h-3.5"></div>
-        <div className="rounded-full bg-green-500 w-3.5 h-3.5"></div>
-      </div>
-      <div className="text-center">bash</div>
-    </div>
   );
 
   const TerminalBodyHistory = useCallback(() => {
